Moved error class/enum to app.h file
