+++ firmware/projects/Demo/CANErrors/main.cc
@@ -2,6 +2,8 @@
 /// @date 2024-11-05
 
 #include <cstdint>
+
+#include "app.h"
 #include "bindings.h"
 #include "shared/comms/can/can_bus.h"
 #include "generated/can/error_msg_registry.h"
@@ -19,23 +21,23 @@ int main(void) {
     uint32_t tick_duration = 100;
 
     // Initialize the error object that sets/sends each error
-    bindings::ErrorHandler error_msg{};
+    ErrorHandler error_msg{};
 
     while (true) {
         error_can_bus.Update();
 
-        // Iterate through each error, setting it and sending it
+        // Test #1: Iterate through each error, setting it and sending it
         for (int i = 0; i < 63; i++) {
             error_msg.setError(i);
             error_msg.sendError(error_can_bus);
         }
 
-        // Set random errors using the enum
-        error_msg.setError(bindings::Error0);
-        error_msg.setError(bindings::Error5);
-        error_msg.setError(bindings::Error10);
-        error_msg.setError(bindings::Error15);
-        error_msg.setError(bindings::Error20);
+        // Test #2: Set random errors using the enum
+        error_msg.setError(Error0);
+        error_msg.setError(Error5);
+        error_msg.setError(Error10);
+        error_msg.setError(Error15);
+        error_msg.setError(Error20);
         error_msg.sendError(error_can_bus);
 
         // Wait for 100ms before repeating the process
